requirements.txt
torch
torchvision
numpy
pandas
streamlit
Pillow
opencv-python